requirements.txt
streamlit
python-dotenv
langchain-openai
langchain-community
openai
pandas
whisper
pyaudio
pydub